refactor: reorganize schema definitions; enhance documentation and structure for clarity
