+++ packages/@studiocms/wysiwyg/src/common/editor-utils.ts
@@ -21,37 +21,6 @@ export const generateHTML = async (editor: Editor): Promise<string> => {
 };
 
 /**
- * Adds an 'inline' storage provider to the given editor instance.
- *
- * This storage provider allows loading and storing project data directly within the editor context.
- * When storing, it serializes the provided data along with the current editor HTML content
- * (generated by `generateHTML(editor)`) into the specified `pageContent` element's `innerText`.
- *
- * @param editor - The editor instance to which the storage provider will be added.
- * @param opts - Options for storage:
- *   - `projectData`: The project data to be loaded.
- *   - `pageContent`: The HTML element where serialized data will be stored.
- * 
- * @deprecated Use `StudioCMSDbStorageAdapter` for database-backed storage instead.
- */
-export const inlineStorage = (
-	editor: Editor,
-	opts: { projectData: ProjectData; pageContent: HTMLElement }
-) => {
-	editor.Storage.add('inline', {
-		async load() {
-			return opts.projectData;
-		},
-		async store(data) {
-			opts.pageContent.innerText = JSON.stringify({
-				...data,
-				__STUDIOCMS_HTML: await generateHTML(editor),
-			});
-		},
-	});
-};
-
-/**
  * Registers a custom storage adapter named 'db' for the provided editor instance.
  * This adapter enables loading and storing project data to a backend database using fetch requests.
  *
